RavenDB-11310 - Move RachisMergedCommand to new file
